+++ crates/var/src/lib.rs
@@ -13,6 +13,7 @@
 
 mod binary;
 mod cbor;
+mod diff;
 pub mod encode;
 mod error;
 mod flyweight;
@@ -28,6 +29,7 @@ mod variant;
 
 pub use binary::Binary;
 pub use cbor::{CborVarError, decode_var_cbor, encode_var_cbor};
+pub use diff::{ValueDiffOptions, value_diff, value_diff3};
 pub use error::{Error, ErrorCode, ErrorCode::*};
 pub use flyweight::Flyweight;
 pub use lambda::Lambda;